De-document USE_SREGEX and error out if it's disabled.
regexec() and friends have been a defined part of libc in POSIX since
sometime in the 90's, and the only current or even non-current system
I can find that _might_ lack it is Windows.  And it's probably
available in some other library on Win if we need to build there;
figure that out when it comes up.