Convert TitleJustification elements to their own enum type, and fixup parsing.
This has the minor side effect that "undef", while previous an
accepted value for TitleJustification and IconJustification, now gives
an error on parsing.  I consider this acceptable, as it was never
documented as a valid value, and if it were used, it would have very
bad side effects any place it was used, so it's never a sensible
choice anyway.