XmbTextExtents() is documented to permit NULL for the rectangles we don't care about, so do that for the ones we're not using the return of.
This is slightly tricky in the case where we're (completely bogusly)
passing the same XRectangle to both params.  I somewhat arbitrarily
decide that we're intending to use it as the one it's named as.  It's
only _somewhat_ arbitrary as a few moments thinking on how it's used
seems to bear that out.