Correct outline when 3D borders not used, but BorderSize is
Without that, right/bottom edge detection is wrong by BorderSize
pixels when moving the outline.

It seems that bw param of MoveOutline() should always be 0 to draw
correct outline, but do not touch it. Inner lines do not touch the
external outline, which is not the case in 3D border case.


git: 30de383741de324cb0ebc1570a35d3657c7c968d