Implement EWMH _NET_WM_VISIBLE{,_ICON}_NAME, telling external EWMH-talking utils what we're calling a window, if it's not one of the standard properties.
I can't immediately find any easily-testable pagers that look at this,
and wmctrm(1) doesn't either, so it's only testing insofar as I can
see that it's setting the properties.  There's also the question of
character encoding, which I'm gleefully ignoring at the moment; see
comments about handling in the future.