Redo figuring feature test macros to be more generic.
Previous code was a hardcoded "Does this look like glibc?  Then it
needs this stuff added".  However, current Illumos (at least with gcc)
also needs some defs added.  So rework it to be more symptomatic; take
the funcs we generally see falling into the cracks, and see which
extra defs may be needed for them.