Bypass BorderedLayout for strut'd windows.
This is a hack, and does the wrong thing.  However, due to the way we
currently handle struts, the previous behavior would force a window to
move outside its own strut, which is nonsense.  A proper implementation
would require evaluating struts individually, not cramming them into
BorderedLayout (which already means something similar but distinct), and
encapsulating those checks in all the places that do window positioning.

That's a major change though; this at least avoids the extra-dumb
behavior, and works no worse than 4.0.3 and pre-r644 code for those
corner cases.